+++ packages/rnative/src/components/elements/datePicker.tsx
@@ -1,6 +1,13 @@
 import React, {useMemo} from 'react';
 
-import {Pressable, StyleSheet, Text, TextStyle, View} from 'react-native';
+import {
+  Pressable,
+  PressableProps,
+  StyleSheet,
+  Text,
+  TextStyle,
+  View,
+} from 'react-native';
 import {colors} from '../../themes/colors';
 import {spacing} from '../../themes/spacing';
 import {textSize} from '../../themes/textSize';
@@ -27,6 +34,7 @@ const DatePicker = ({
   format,
   value,
   onChange,
+  disabled,
 }: DatePickerProps) => {
   const {onOpen, isOpen, onClose} = useDisclose();
 
@@ -47,6 +55,16 @@ const DatePicker = ({
     }
   }, [value, format, placeholder]);
 
+  const customStyle = useMemo<PressableProps['style']>(() => {
+    let additionalStyle: PressableProps['style'] = {};
+
+    if (disabled) {
+      additionalStyle = {...additionalStyle, backgroundColor: colors.muted};
+    }
+
+    return {...style.DropdownButton, ...additionalStyle};
+  }, [disabled]);
+
   const customTextStyle = useMemo<TextStyle>(() => {
     let additionalStyle: TextStyle = {};
 
@@ -54,15 +72,20 @@ const DatePicker = ({
       additionalStyle = {...additionalStyle, color: colors.mutedForeground};
     }
 
+    if (disabled) {
+      additionalStyle = {...additionalStyle, color: colors.mutedForeground};
+    }
+
     return {...style.DropdownButtonText, ...additionalStyle};
-  }, [display, placeholder]);
+  }, [display, placeholder, disabled]);
 
   return (
     <View>
       <Pressable
-        style={style.DropdownButton}
+        style={customStyle}
         android_ripple={{color: colors.accent}}
-        onPress={onOpen}>
+        onPress={onOpen}
+        disabled={disabled}>
         <Icon
           name="calendar-month"
           size={textSize.sm}