+++ packages/rnative/src/components/ui/input.tsx
@@ -1,45 +1,45 @@
-import {useColors} from '../../hooks/useColors';
-import {useSpacing} from '../../hooks/useSpacing';
-import {useTextSize} from '../../hooks/useTextSize';
 import * as React from 'react';
 import {StyleSheet, TextInput, TextInputProps} from 'react-native';
 import {rem} from '../../utils/helper';
+import {textSize} from '../../themes/textSize';
+import {colors} from '../../themes/colors';
+import {spacing} from '../../themes/spacing';
 
 export interface InputProps extends Omit<TextInputProps, 'onChange'> {
   onChange?: TextInputProps['onChangeText'];
+  disabled?: boolean;
 }
 
 const Input = React.forwardRef<TextInput, InputProps>(
-  ({onChange, onChangeText, ...props}, ref) => {
+  ({onChange, onChangeText, disabled, readOnly, ...props}, ref) => {
     const onChangeRef = React.useRef<InputProps['onChange']>();
     const onChangeTextRef = React.useRef<InputProps['onChangeText']>();
 
-    const colors = useColors();
-    const textSize = useTextSize();
-    const spacing = useSpacing();
+    const style = React.useMemo(() => {
+      let additionalStyle: TextInputProps['style'] = {};
 
-    const style = React.useMemo(
-      () =>
-        StyleSheet.create({
-          Input: {
-            fontSize: textSize.sm,
-            color: colors.foreground,
-            borderWidth: 1,
-            borderColor: colors.input,
-            paddingHorizontal: rem`0.75`,
-            paddingVertical: rem`0.5`,
-            borderRadius: spacing.radius,
-            ...((props.style as object) ?? {}),
-          },
-        }),
-      [
-        colors.foreground,
-        colors.input,
-        props.style,
-        spacing.radius,
-        textSize.sm,
-      ],
-    );
+      if (disabled) {
+        additionalStyle = {
+          ...additionalStyle,
+          backgroundColor: colors.muted,
+          color: colors.mutedForeground,
+        };
+      }
+
+      return StyleSheet.create({
+        Input: {
+          fontSize: textSize.sm,
+          color: colors.foreground,
+          borderWidth: 1,
+          borderColor: colors.input,
+          paddingHorizontal: rem`0.75`,
+          paddingVertical: rem`0.5`,
+          borderRadius: spacing.radius,
+          ...additionalStyle,
+          ...((props.style as object) ?? {}),
+        },
+      });
+    }, [props.style, disabled]);
 
     React.useEffect(() => {
       if (!onChange) {
@@ -73,6 +73,7 @@ const Input = React.forwardRef<TextInput, InputProps>(
         onChangeText={handleChangeText}
         style={style.Input}
         placeholderTextColor={colors.mutedForeground}
+        readOnly={readOnly || disabled}
       />
     );
   },